+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.6.0
+
+### Added
+
+- Support to download Sentinel-2 products from Google Cloud storage.
+- Added support to download by text match.
+- Added `download by-geometry` and `download by-title` to Fire commands.
+
+
 ## Released
 
 ## 0.5.1